renamed files, added aliases, restructured directories then fixed imports
