+++ TTS/babel.config.js
@@ -2,5 +2,19 @@ module.exports = function(api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
+    plugins: [
+      [
+        'module-resolver',
+        {
+          root: ['./src/'],
+          alias: {
+            '@components': './src/components',
+            '@contexts': './src/contexts',
+            '@hooks': './src/hooks',
+            '@services': './src/services',
+          },
+        },
+      ],
+    ],
   };
 };
