+++ TTS/src/__tests__/RiskNote.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react-native';
-import RiskNote from '../components/risk_form/RiskNote';
+import RiskNote from '@components/risk_form/RiskNote';
 
 describe('<RiskNote />', () => {
   const mockOnChange = jest.fn();
