+++ TTS/src/components/Main.jsx
@@ -1,12 +1,12 @@
 import { StyleSheet, View, Text } from 'react-native';
 import React, { useEffect, useContext } from 'react';
 import { Route, Routes, Navigate } from 'react-router-native';
-import ProjectList from './ProjectList';
-import WorkSafetyForm from './risk_form/WorkSafetyForm';
-import SignIn from './SignIn';
-import AppBar from './AppBar';
+import ProjectList from '@components/project-list/ProjectList';
+import WorkSafetyForm from '@components/risk-form/RiskForm';
+import SignIn from '@components/sign-in/SignIn';
+import AppBar from '@components/app-bar/AppBar';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { UserContext } from '../contexts/UserContext';
+import { UserContext } from '@contexts/UserContext';
 
 const Main = () => {
   const { username, setUsername } = useContext(UserContext)
@@ -34,7 +34,7 @@ const Main = () => {
         <Routes>
           <Route path='/' element={
             <>
-              <ProjectList /> 
+              <ProjectList/> 
             </>
           }
           />
